migrated assertions into separate component
