+++ src/components/EndpointTestComponent/Endpoint.jsx
@@ -3,7 +3,7 @@ import { Draggable } from 'react-beautiful-dnd';
 import styles from './Endpoint.module.scss';
 import style from '../ReactTestComponent/Render/Render.module.scss';
 import styled from '../ReactTestComponent/Render/Prop.module.scss';
-import stylez from '../ReactTestComponent/Assertion/Assertion.module.scss';
+import EndpointAssertion from './EndpointAssertion';
 
 import {
   deleteEndpoint,
@@ -17,44 +17,8 @@ import {
 const closeIcon = require('../../assets/images/close.png');
 const dragIcon = require('../../assets/images/drag-vertical.png');
 const minusIcon = require('../../assets/images/minus-box-outline.png');
-const questionIcon = require('../../assets/images/help-circle.png');
 
 const Endpoint = ({ endpoint, index, dispatchToEndpointTestCase }) => {
-  const jestMatchers = [
-    'to Be',
-    'to Equal (object)',
-    'to Have Been Called',
-    'to Have Been Called Times (number)',
-    'to Have Been Called With (arg1,...)',
-    'to Have Been Last Called With (arg1,...)',
-    'to Have Been Nth Called With (nth call, arg1,...)',
-    'to Have Length (number)',
-    'to Have Property (keyPath, value[optional])',
-    'to Be Close To (number, number of digits[optional])',
-    'to Be Defined',
-    'to Be Undefined',
-    'to Be Falsy',
-    'to Be Truthy',
-    'to Be NaN',
-    'to Be Greater Than (number)',
-    'to Be Greater Than Or Equal (number)',
-    'to Be Less Than (number)',
-    'to Be Less Than Or Equal (number)',
-    'to Be Instance Of (Class)',
-    'to Contain (item in an array)',
-    'to Contain Equal (an object in an array)',
-    'to Match (regexp or string)',
-    'to Match Object (object)',
-    'to Srict Equal (object)',
-    'to Throw (error[optional])',
-  ];
-
-  //for mock fuctions only:
-  //   'to Have Returned',
-  //   'to Have Returned __ Times (number)',
-  //   'to Have Last Returned With',
-  // ];
-
   const handleChangeEndpointFields = (e, field) => {
     let updatedEndpoint = { ...endpoint };
 
@@ -172,79 +136,17 @@ const Endpoint = ({ endpoint, index, dispatchToEndpointTestCase }) => {
                   />
                 </div>
               </div>
-            </div>
-            {/* ------------------------------------------------------------------------      */}
-            <div id={styles.groupFlexboxAssertion}>
-              <div id={styles.labelInput}>
-                <label htmlFor='requestBody'>Expect Response</label>
-                <div id={styles.inputFlexBox}>
-                  <input
-                    type='text'
-                    list='responseProperties'
-                    value={endpoint.expectedResponse}
-                    onChange={(e) => handleChangeEndpointFields(e, 'expectedResponse')}
-                  />
-                  <datalist id='responseProperties'>
-                    <option value='Headers'></option>
-                    <option value='Status'></option>
-                    <option value='Body'></option>
-                    <option value='Message'></option>
-                    <option value='Length'></option>
-                  </datalist>
-                </div>
-              </div>
-              <div id={styles.dropdownWrapper}>
-                {/* <label htmlFor='value'>Assertion</label> */}
-                <div id={stylez.matcherLabelFlexBox}>
-                  <div>
-                    <label htmlFor='matcher'>Matcher</label>
-                  </div>
-                  <div>
-                    Not?
-                    <input
-                      type='checkbox'
-                      // checked={statement.isNot}
-                      // onChange={(e) => handleChangeAssertionFields(e, 'isNot')}
-                    />
-                  </div>
-                </div>
-                {/* -------------------------------------------------------------------------------------------------- */}
-
-                <div id={styles.dropdownFlex}>
-                  <select
-                    id='method'
-                    value={endpoint.assertion}
-                    onChange={(e) => handleChangeEndpointFields(e, 'assertion')}
-                  >
-                    {jestMatchers.map((assertion) => (
-                      <option value={assertion}>{assertion}</option>
-                    ))}
-                  </select>{' '}
-                  <span id={stylez.hastooltip} role='tooltip'>
-                    <img src={questionIcon} alt='help' />
-                    <span id={stylez.tooltip}>
-                      {/* <ToolTipMatcher toolTipType={statement.matcherType} /> */}
-                    </span>
-                  </span>
-                </div>
-                {/* <div id={stylez.autoTool}>
-                  <input type='text' /> */}
-
-                {/* </div> */}
-              </div>
-              <div id={styles.labelInput}>
-                <label htmlFor='value'>Expected Value</label>
-                <div id={styles.inputFlexBox}>
-                  <input
-                    type='text'
-                    name='value'
-                    placeholder='eg. 200'
-                    onChange={(e) => handleChangeEndpointFields(e, 'value')}
-                  />
-                </div>
-              </div>
             </div>{' '}
-            {/* -------------------------------------------------------------------------------------------------- */}
+            {endpoint.assertions.map((assertion, i) => {
+              return (
+                <EndpointAssertion
+                  endpoint={endpoint}
+                  index={index}
+                  id={i}
+                  handleChangeEndpointFields={handleChangeEndpointFields}
+                />
+              );
+            })}{' '}
             {endpoint.post && (
               <div id={style.RenderContainer} style={{ margin: '10px 0 0 0' }}>
                 <label htmlFor='Header' id={styles.labelInputPost}>
